2-7616273326827507494/testbed/requirements.txt
langchain>=1.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=1.0.0
langchain-core>=0.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pytest>=8.0.0
rank-bm25>=0.2.2
